Add kmeans templated class
Adds a kmeans templated class. Hopefully can do simple clustering as is.
Needs a type T that can overrides the + operator with two T types and / operator with a T and scalar type. A distance function must be passed into the object's () operator, and can be done with a lambda.