Add backward one step and change point size/shape
Adds the ability to go backward one step. It does not save more than one
step at a time, and when it's possible to go back, the button will
enable. Otherwise, it will disable.

Adds the ability to change point size and shape. The graph will update
if it's not playing the animation, since it will update on the next step
automatically anyway.