Fix backward step bug
Fixes going backward a step. Since there's a stop condition when
assignments don't change, it was triggering when going backward a step.
This adds a flag for k-means to ignore a step's assignment stop
condition.

Also fixes the visual bug of centroids not updating with going back a step for the 3D case.